requirements.txt
pyowm
pprintpp
simplejson
requests
softlayer-object-storage